+++ server/api/src/main/java/com/walmartlabs/concord/server/api/process/ProcessEntry.java
@@ -16,6 +16,8 @@ public class ProcessEntry implements Serializable {
     private final UUID instanceId;
     private final ProcessKind kind;
     private final UUID parentInstanceId;
+    private final UUID orgId;
+    private final String orgName;
     private final UUID projectId;
     private final String projectName;
     private final Date createdAt;
@@ -30,6 +32,8 @@ public class ProcessEntry implements Serializable {
     public ProcessEntry(@JsonProperty("instanceId") UUID instanceId,
                         @JsonProperty("kind") ProcessKind kind,
                         @JsonProperty("parentInstanceId") UUID parentInstanceId,
+                        @JsonProperty("orgId") UUID orgId,
+                        @JsonProperty("orgName") String orgName,
                         @JsonProperty("projectId") UUID projectId,
                         @JsonProperty("projectName") String projectName,
                         @JsonProperty("createdAt") Date createdAt,
@@ -42,6 +46,8 @@ public class ProcessEntry implements Serializable {
         this.instanceId = instanceId;
         this.kind = kind;
         this.parentInstanceId = parentInstanceId;
+        this.orgId = orgId;
+        this.orgName = orgName;
         this.projectId = projectId;
         this.projectName = projectName;
         this.createdAt = createdAt;
@@ -66,6 +72,14 @@ public class ProcessEntry implements Serializable {
 
     public UUID getParentInstanceId() {
         return parentInstanceId;
+    }
+
+    public UUID getOrgId() {
+        return orgId;
+    }
+
+    public String getOrgName() {
+        return orgName;
     }
 
     public UUID getProjectId() {
@@ -110,6 +124,8 @@ public class ProcessEntry implements Serializable {
                 "instanceId=" + instanceId +
                 ", kind=" + kind +
                 ", parentInstanceId=" + parentInstanceId +
+                ", orgId=" + orgId +
+                ", orgName='" + orgName + '\'' +
                 ", projectId=" + projectId +
                 ", projectName='" + projectName + '\'' +
                 ", createdAt=" + createdAt +